requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
defusedxml==0.7.1
Django==3.2.9
django-environ==0.8.1
django-filter==21.1
django-neomodel==0.0.7
djangorestframework==3.12.4
google-api-core==2.3.0
google-auth==2.3.3
google-cloud-secret-manager==2.8.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
httpie==2.6.0
idna==3.3
importlib-metadata==4.8.2
libcst==0.3.23
Markdown==3.3.6
mypy-extensions==0.4.3
neo4j==4.4.0
neo4j-driver==4.3.6
neobolt==1.7.17
neomodel==4.0.7
proto-plus==1.19.8
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
PySocks==1.7.1
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.8
Shapely==1.7.1
six==1.16.0
sqlparse==0.4.2
typing-inspect==0.7.1
typing_extensions==4.0.1
urllib3==1.26.7
zipp==3.6.0
